requirements.txt
asgiref==3.9.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
dj==0.1
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
idna==3.10
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
requests==2.32.4
sqlparse==0.5.3
urllib3==2.5.0
python-decouple==3.8